Make period correspondences

make_period_correspondences() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make period correspondences. The function returns Correspondences (an output object of multiple potential types).

Usage

make_period_correspondences(
  descriptions_chr,
  integers_1L_lgl = TRUE,
  name_1L_chr = "period",
  plural_chr = "s",
  range_int = 1L:12L,
  reference_1L_int = integer(0),
  spaced_1L_lgl = TRUE,
  type_1L_chr = "descriptions",
  units_chr = c("minute", "hour", "week", "month", "year")
)

Arguments

descriptions_chr

Descriptions (a character vector)

integers_1L_lgl

Integers (a logical vector of length one), Default: TRUE

name_1L_chr

Name (a character vector of length one), Default: 'period'

plural_chr

Plural (a character vector), Default: 's'

range_int

Range (an integer vector), Default: 1L:12L

reference_1L_int

Reference (an integer vector of length one), Default: integer(0)

spaced_1L_lgl

Spaced (a logical vector of length one), Default: TRUE

type_1L_chr

Type (a character vector of length one), Default: 'descriptions'

units_chr

Units (a character vector), Default: c("minute", "hour", "week", "month", "year")

Value

Correspondences (an output object of multiple potential types)