Manufacture a new object

manufacture.ready4use_dataverses() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of files to be imported from a dataverse. The function returns Read (a list of functions).

manufacture.ready4use_imports() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of sources of raw (un-processed) data to import. The function is called for its side effects and does not return a value.

manufacture method applied to Ready4useDyad

Usage

manufacture.ready4use_dataverses(x, type_1L_chr = "read_fn")

# S4 method for class 'ready4use_dataverses'
manufacture(x, type_1L_chr = "read_fn")

manufacture.ready4use_imports(
  x,
  forced_choice_chr = NA_character_,
  script_args_ls = NULL
)

# S4 method for class 'ready4use_imports'
manufacture(x, forced_choice_chr = NA_character_, script_args_ls = NULL)

# S4 method for class 'Ready4useDyad'
manufacture(
  x,
  correspondences_r3 = ready4show::ready4show_correspondences(),
  dummys_ls = list(),
  indcs_int = 1L,
  flatten_1L_lgl = F,
  match_1L_chr = character(0),
  restrict_to_chr = character(0),
  remove_ctg_1L_lgl = F,
  target_1L_chr = "var_ctg_chr",
  type_1L_chr = "all",
  what_1L_chr = "factors",
  ...
)

Arguments

x

An object of class Ready4useDyad

type_1L_chr

Type (a character vector of length one), Default: 'all'

forced_choice_chr

Forced choice (a character vector), Default: 'NA'

script_args_ls

Script arguments (a list), Default: NULL

correspondences_r3

Correspondences (a ready4 submodule), Default: ready4show::ready4show_correspondences()

dummys_ls

Dummys (a list), Default: list()

indcs_int

Indices (an integer vector), Default: 1

flatten_1L_lgl

Flatten (a logical vector of length one), Default: F

match_1L_chr

Match (a character vector of length one), Default: character(0)

restrict_to_chr

Restrict to (a character vector), Default: character(0)

remove_ctg_1L_lgl

Remove category (a logical vector of length one), Default: F

target_1L_chr

Target (a character vector of length one), Default: 'var_ctg_chr'

what_1L_chr

What (a character vector of length one), Default: 'factors'

...

Additional arguments

Value

Read (a list of functions)

No return value, called for side effects.

Object (an output object of multiple potential types)