Procure data from a model module

procure.ready4use_dataverses() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of files to be imported from a dataverse. The function returns Data (an output object of multiple potential types).

procure.ready4use_imports() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of sources of raw (un-processed) data to import. The function is called for its side effects and does not return a value.

procure method applied to Ready4useIngest

Usage

procure.ready4use_dataverses(
  x,
  save_dir_path_1L_chr = "",
  unlink_1L_lgl = T,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  key_1L_chr = Sys.getenv("DATAVERSE_KEY")
)

# S4 method for class 'ready4use_dataverses'
procure(
  x,
  save_dir_path_1L_chr = "",
  unlink_1L_lgl = T,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  key_1L_chr = Sys.getenv("DATAVERSE_KEY")
)

procure.ready4use_imports(
  x,
  inc_script_1L_lgl = T,
  forced_choice_chr = NA_character_
)

# S4 method for class 'ready4use_imports'
procure(x, inc_script_1L_lgl = T, forced_choice_chr = NA_character_)

# S4 method for class 'Ready4useIngest'
procure(x, fl_nm_1L_chr = NA_character_)

Arguments

x

An object of class Ready4useIngest

save_dir_path_1L_chr

Save directory path (a character vector of length one), Default: ”

unlink_1L_lgl

Unlink (a logical vector of length one), Default: T

server_1L_chr

Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")

key_1L_chr

Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")

inc_script_1L_lgl

Include script (a logical vector of length one), Default: T

forced_choice_chr

Forced choice (a character vector), Default: 'NA'

fl_nm_1L_chr

File name (a character vector of length one), Default: 'NA'

Value

Data (an output object of multiple potential types)

No return value, called for side effects.

Object (an output object of multiple potential types)