ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.

Create a new valid instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.

Usage

ready4use_mapes(x = make_pt_ready4use_mapes())

make_pt_ready4use_mapes(
  param_name_chr = character(0),
  sex_age_band_chr = character(0),
  mape_05_yr_mde_dbl = numeric(0),
  mape_10_yr_mde_dbl = numeric(0),
  mape_15_yr_mde_dbl = numeric(0),
  mape_05_yr_min_dbl = numeric(0),
  mape_10_yr_min_dbl = numeric(0),
  mape_15_yr_min_dbl = numeric(0),
  mape_05_yr_max_dbl = numeric(0),
  mape_10_yr_max_dbl = numeric(0),
  mape_15_yr_max_dbl = numeric(0),
  mape_05_yr_shp_dbl = numeric(0),
  mape_10_yr_shp_dbl = numeric(0),
  mape_15_yr_shp_dbl = numeric(0)
)

is_ready4use_mapes(x)

Arguments

x

An object of any type

param_name_chr

Parameter name (a character vector), Default: character(0)

sex_age_band_chr

Sex age band (a character vector), Default: character(0)

mape_05_yr_mde_dbl

Mean absolute prediction error 05 year mode (a double vector), Default: numeric(0)

mape_10_yr_mde_dbl

Mean absolute prediction error 10 year mode (a double vector), Default: numeric(0)

mape_15_yr_mde_dbl

Mean absolute prediction error 15 year mode (a double vector), Default: numeric(0)

mape_05_yr_min_dbl

Mean absolute prediction error 05 year minimum (a double vector), Default: numeric(0)

mape_10_yr_min_dbl

Mean absolute prediction error 10 year minimum (a double vector), Default: numeric(0)

mape_15_yr_min_dbl

Mean absolute prediction error 15 year minimum (a double vector), Default: numeric(0)

mape_05_yr_max_dbl

Mean absolute prediction error 05 year maximum (a double vector), Default: numeric(0)

mape_10_yr_max_dbl

Mean absolute prediction error 10 year maximum (a double vector), Default: numeric(0)

mape_15_yr_max_dbl

Mean absolute prediction error 15 year maximum (a double vector), Default: numeric(0)

mape_05_yr_shp_dbl

Mean absolute prediction error 05 year shape (a double vector), Default: numeric(0)

mape_10_yr_shp_dbl

Mean absolute prediction error 10 year shape (a double vector), Default: numeric(0)

mape_15_yr_shp_dbl

Mean absolute prediction error 15 year shape (a double vector), Default: numeric(0)

Value

A validated instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.

A prototype for ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.

A logical value, TRUE if a valid instance of the ready4 S3 class for tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.

Details

Tibble object that stores simulation structural parameters relating to Mean Absolute Prediction Errors.