Renew (update) values

renew.ready4use_dictionary() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the ready4 s3 class defining a data dictionary tibble. The function is called for its side effects and does not return a value.

renew.ready4use_imports() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the ready4 submodule class for tibble object lookup table of sources of raw (un-processed) data to import. The function is called for its side effects and does not return a value.

renew method applied to Ready4useDyad

Usage

renew.ready4use_dictionary(
  x,
  var_nm_chr = NA_character_,
  var_ctg_chr = NA_character_,
  var_desc_chr = NA_character_,
  var_type_chr = NA_character_,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  new_ready4_dict_r3 = deprecated(),
  slice_indcs_int = NA_integer_
)

# S4 method for class 'ready4use_dictionary'
renew(
  x,
  var_nm_chr = NA_character_,
  var_ctg_chr = NA_character_,
  var_desc_chr = NA_character_,
  var_type_chr = NA_character_,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  new_ready4_dict_r3 = deprecated(),
  slice_indcs_int = NA_integer_
)

renew.ready4use_imports(
  x,
  local_file_src_chr = NA_character_,
  path_to_make_script_chr = NA_character_,
  download_url_chr = NA_character_,
  inc_file_main_chr = NA_character_,
  inc_fls_to_rename_ls = list(),
  new_nms_for_inc_fls_ls = list(),
  filter_cdn_1L_chr = NA_character_,
  local_to_url_vec_chr = NA_character_,
  slice_indcs_int = NA_integer_,
  urls_vec_chr = NA_character_
)

# S4 method for class 'ready4use_imports'
renew(
  x,
  local_file_src_chr = NA_character_,
  path_to_make_script_chr = NA_character_,
  download_url_chr = NA_character_,
  inc_file_main_chr = NA_character_,
  inc_fls_to_rename_ls = list(),
  new_nms_for_inc_fls_ls = list(),
  filter_cdn_1L_chr = NA_character_,
  local_to_url_vec_chr = NA_character_,
  slice_indcs_int = NA_integer_,
  urls_vec_chr = NA_character_
)

# S4 method for class 'Ready4useDyad'
renew(
  x,
  arrange_by_1L_chr = c("category", "name", "both"),
  categories_chr = character(0),
  drop_chr = character(0),
  dictionary_lups_ls = list(),
  dictionary_r3 = ready4use_dictionary(),
  dummys_ls = NULL,
  exclude_chr = character(0),
  factors_chr = character(0),
  dyad_ls = NULL,
  fn = NULL,
  fn_args_ls = NULL,
  lup_tb = NULL,
  match_var_nm_1L_chr = character(0),
  method_1L_chr = c("first", "sample"),
  names_chr = character(0),
  new_val_xx = NULL,
  remove_old_lbls_1L_lgl = T,
  tfmn_1L_chr = "capitalise",
  type_1L_chr = c("label", "base", "case", "drop", "dummys", "join", "keep", "levels",
    "mutate", "new", "rbind", "unlabel", "update", "sequential", "batch", "self"),
  uid_var_nm_1L_chr = character(0),
  var_ctg_chr = "Uncategorised",
  vars_chr = character(0),
  what_1L_chr = c("all", "dataset", "dictionary"),
  ...
)

Arguments

x

An object of class Ready4useDyad

var_nm_chr

Variable name (a character vector), Default: 'NA'

var_ctg_chr

Variable category (a character vector), Default: 'Uncategorised'

var_desc_chr

Variable description (a character vector), Default: 'NA'

var_type_chr

Variable type (a character vector), Default: 'NA'

filter_cdn_1L_chr

Filter condition (a character vector of length one), Default: 'NA'

new_cases_r3

New cases (a ready4 submodule), Default: NULL

new_ready4_dict_r3

New ready4 dictionary (a ready4 submodule), Default: deprecated()

slice_indcs_int

Slice indices (an integer vector), Default: NA

local_file_src_chr

Local file source (a character vector), Default: 'NA'

path_to_make_script_chr

Path to make script (a character vector), Default: 'NA'

download_url_chr

Download url (a character vector), Default: 'NA'

inc_file_main_chr

Include file main (a character vector), Default: 'NA'

inc_fls_to_rename_ls

Include files to rename (a list), Default: list()

new_nms_for_inc_fls_ls

New names for include files (a list), Default: list()

local_to_url_vec_chr

Local to url vector (a character vector), Default: 'NA'

urls_vec_chr

Urls vector (a character vector), Default: 'NA'

arrange_by_1L_chr

Arrange by (a character vector of length one), Default: c("category", "name", "both")

categories_chr

Categories (a character vector), Default: character(0)

drop_chr

Drop (a character vector), Default: character(0)

dictionary_lups_ls

Dictionary lookup tables (a list), Default: list()

dictionary_r3

Dictionary (a ready4 submodule), Default: ready4use_dictionary()

dummys_ls

Dummys (a list), Default: NULL

exclude_chr

Exclude (a character vector), Default: character(0)

factors_chr

Factors (a character vector), Default: character(0)

dyad_ls

Dyad (a list), Default: NULL

fn

Function (a function), Default: NULL

fn_args_ls

Function arguments (a list), Default: NULL

lup_tb

Lookup table (a tibble), Default: NULL

match_var_nm_1L_chr

Match variable name (a character vector of length one), Default: character(0)

method_1L_chr

Method (a character vector of length one), Default: c("first", "sample")

names_chr

Names (a character vector), Default: character(0)

new_val_xx

New value (an output object of multiple potential types), Default: NULL

remove_old_lbls_1L_lgl

Remove old labels (a logical vector of length one), Default: T

tfmn_1L_chr

Transformation (a character vector of length one), Default: 'capitalise'

type_1L_chr

Type (a character vector of length one), Default: c("label", "base", "case", "drop", "dummys", "join", "keep", "levels", "mutate", "new", "rbind", "unlabel", "update", "sequential", "batch", "self")

uid_var_nm_1L_chr

Unique identifier variable name (a character vector of length one), Default: character(0)

vars_chr

Variables (a character vector), Default: character(0)

what_1L_chr

What (a character vector of length one), Default: c("all", "dataset", "dictionary")

...

Additional arguments

Value

x (An object)

x (An object)

x (An object of class Ready4useDyad)