Update column names

update_column_names() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update column names. The function is called for its side effects and does not return a value.

Usage

update_column_names(
  X_Ready4useDyad,
  patterns_ls = list(c("[[:space:]]", "")),
  update_desc_1L_lgl = FALSE
)

Arguments

X_Ready4useDyad

PARAM_DESCRIPTION

patterns_ls

Patterns (a list), Default: list(c("[:space:]", ""))

update_desc_1L_lgl

Update description (a logical vector of length one), Default: FALSE

Value

X (A dataset and data dictionary pair.)