Update correspondences

update_correspondences() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update correspondences. The function returns Correspondences (a list).

Usage

update_correspondences(
  correspondences_ls,
  dyad_ls = NULL,
  new_ls = NULL,
  filter_cdn_1L_chr = "!is.na(new_nms_chr)",
  range_int = 1L:12L,
  reference_1L_int = 2L,
  spaced_1L_lgl = TRUE,
  type_1L_chr = "sequence",
  units_chr = c("minute", "hour", "week", "month", "year")
)

Arguments

correspondences_ls

Correspondences (a list)

dyad_ls

Dyad (a list), Default: NULL

new_ls

New (a list), Default: NULL

filter_cdn_1L_chr

Filter condition (a character vector of length one), Default: '!is.na(new_nms_chr)'

range_int

Range (an integer vector), Default: 1L:12L

reference_1L_int

Reference (an integer vector of length one), Default: 2

spaced_1L_lgl

Spaced (a logical vector of length one), Default: TRUE

type_1L_chr

Type (a character vector of length one), Default: 'sequence'

units_chr

Units (a character vector), Default: c("minute", "hour", "week", "month", "year")

Value

Correspondences (a list)