Update dyad list

update_dyad_ls() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update dyad list. The function returns Dyad (a list).

Usage

update_dyad_ls(
  dyad_ls,
  add_lups_1L_lgl = F,
  arrange_1L_chr = c("var_ctg_chr, var_nm_chr"),
  factors_chr = character(0),
  range_int = 1L:12L,
  recode_ls = NULL,
  reference_1L_int = 2L,
  spaced_1L_lgl = TRUE,
  standard_spaces_1L_lgl = F,
  tfmn_cls_1L_chr = "character",
  tfmns_ls = list(bind = identity, class = as.character),
  type_1L_chr = c("sequence", "composite", "bind", "class", "default", "interval",
    "reference"),
  units_chr = c("minute", "hour", "week", "month", "year"),
  uid_var_nm_1L_chr = character(0)
)

Arguments

dyad_ls

Dyad (a list)

add_lups_1L_lgl

Add lookup tables (a logical vector of length one), Default: F

arrange_1L_chr

Arrange (a character vector of length one), Default: c("var_ctg_chr, var_nm_chr")

factors_chr

Factors (a character vector), Default: character(0)

range_int

Range (an integer vector), Default: 1L:12L

recode_ls

Recode (a list), Default: NULL

reference_1L_int

Reference (an integer vector of length one), Default: 2

spaced_1L_lgl

Spaced (a logical vector of length one), Default: TRUE

standard_spaces_1L_lgl

Standard spaces (a logical vector of length one), Default: F

tfmn_cls_1L_chr

Transformation class (a character vector of length one), Default: 'character'

tfmns_ls

Transformations (a list), Default: list(bind = identity, class = as.character)

type_1L_chr

Type (a character vector of length one), Default: c("sequence", "composite", "bind", "class", "default", "interval", "reference")

units_chr

Units (a character vector), Default: c("minute", "hour", "week", "month", "year")

uid_var_nm_1L_chr

Unique identifier variable name (a character vector of length one), Default: character(0)

Value

Dyad (a list)